second enemy

+++ Assets/Scripts/Bullet.cs
@@ -4,7 +4,7 @@ public class Bullet : MonoBehaviour
 {
     [SerializeField] float speed;
     [SerializeField] float lifetime;
-    public int damage = 25;
+    public float damage = 25;
 
     void Start()
     {
